requirements.txt
fastapi>=0.116.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
aiosqlite>=0.20.0,<1.0.0
httpx>=0.28.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
PyYAML>=6.0.0,<7.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
jinja2>=3.1.0,<4.0.0
playwright>=1.49.0,<2.0.0
python-dotenv>=1.0.1,<2.0.0
